+++ src/utils/utilFunctions.ts
@@ -1,0 +1,12 @@
+export const changeTabName = (tab: string) => {
+  switch (tab) {
+    case "친구":
+      return "friends";
+    case "파티":
+      return "parties";
+    case "길드":
+      return "guilds";
+    default:
+      return "";
+  }
+};
